Cleanup of workspaces and structure.
