--- workspaces/starter.jupyterlab-workspace
@@ -1,1 +1,0 @@
-{"data":{"layout-restorer:data":{"main":{"dock":{"type":"split-area","orientation":"vertical","sizes":[0.5,0.5],"children":[{"type":"tab-area","currentIndex":1,"widgets":["terminal:1","terminal:2"]},{"type":"tab-area","currentIndex":0,"widgets":["notebook:workspaces/Untitled1.ipynb"]}]},"current":"terminal:2"},"down":{"size":0,"widgets":[]},"left":{"collapsed":false,"visible":true,"current":"filebrowser","widgets":["filebrowser","running-sessions","git-sessions","@jupyterlab/toc:plugin","extensionmanager.main-view"],"widgetStates":{"jp-running-sessions":{"sizes":[0.16666666666666666,0.16666666666666666,0.16666666666666666,0.16666666666666666,0.16666666666666666,0.16666666666666666],"expansionStates":[false,false,false,false,false,false]},"extensionmanager.main-view":{"sizes":[0.3333333333333333,0.3333333333333333,0.3333333333333333],"expansionStates":[false,false,false]}}},"right":{"collapsed":true,"visible":true,"widgets":["jp-property-inspector","debugger-sidebar"],"widgetStates":{"jp-debugger-sidebar":{"sizes":[0.2,0.2,0.2,0.2,0.2],"expansionStates":[false,false,false,false,false]}}},"relativeSizes":[0.15991077119184194,0.840089228808158,0],"top":{"simpleVisibility":true}},"docmanager:recents":{"opened":[{"path":"workspaces","contentType":"directory","root":"/home/jovyan"},{"path":"workspaces/Untitled1.ipynb","contentType":"notebook","factory":"Notebook","root":"/home/jovyan"}],"closed":[]},"file-browser-filebrowser:cwd":{"path":"workspaces"},"notebook:workspaces/Untitled1.ipynb":{"data":{"path":"workspaces/Untitled1.ipynb","factory":"Notebook"}},"terminal:1":{"data":{"name":"1"}},"terminal:2":{"data":{"name":"2"}}},"metadata":{"id":"starter","last_modified":"2024-07-17T21:18:11.194040+00:00","created":"2024-07-17T21:18:11.194040+00:00"}}
