--- workspaces/uc-demo.jupyterlab-workspace
@@ -1,1 +1,0 @@
-{"data":{"layout-restorer:data":{"main":{"dock":{"type":"tab-area","currentIndex":4,"widgets":["notebook:workspaces/Untitled.ipynb","terminal:1","terminal:2","terminal:3","terminal:4","terminal:5","terminal:6","terminal:7"]},"current":"terminal:4"},"down":{"size":0,"widgets":[]},"left":{"collapsed":false,"visible":true,"current":"filebrowser","widgets":["filebrowser","running-sessions","git-sessions","@jupyterlab/toc:plugin","extensionmanager.main-view"],"widgetStates":{"jp-running-sessions":{"sizes":[0.16666666666666666,0.16666666666666666,0.16666666666666666,0.16666666666666666,0.16666666666666666,0.16666666666666666],"expansionStates":[false,false,false,false,false,false]},"extensionmanager.main-view":{"sizes":[0.3333333333333333,0.3333333333333333,0.3333333333333333],"expansionStates":[false,false,false]}}},"right":{"collapsed":true,"visible":true,"widgets":["jp-property-inspector","debugger-sidebar"],"widgetStates":{"jp-debugger-sidebar":{"sizes":[0.2,0.2,0.2,0.2,0.2],"expansionStates":[false,false,false,false,false]}}},"relativeSizes":[0.18893072289156626,0.8110692771084337,0],"top":{"simpleVisibility":true}},"docmanager:recents":{"opened":[{"path":"workspaces","contentType":"directory","root":"/home/jovyan"},{"path":"workspaces/Untitled.ipynb","contentType":"notebook","factory":"Notebook","root":"/home/jovyan"},{"path":"workspaces/uc_oss_catalog_viewer.py","contentType":"file","factory":"Editor","root":"/home/jovyan"}],"closed":[{"path":"workspaces/uc_oss_catalog_viewer.py","contentType":"file","factory":"Editor","root":"/home/jovyan"}]},"file-browser-filebrowser:cwd":{"path":"workspaces"},"notebook:workspaces/Untitled.ipynb":{"data":{"path":"workspaces/Untitled.ipynb","factory":"Notebook"}},"workspace-ui:lastSave":"workspaces/uc-demo.jupyterlab-workspace","terminal:1":{"data":{"name":"1"}},"terminal:2":{"data":{"name":"2"}},"terminal:3":{"data":{"name":"3"}},"terminal:4":{"data":{"name":"4"}},"terminal:5":{"data":{"name":"5"}},"terminal:6":{"data":{"name":"6"}},"terminal:7":{"data":{"name":"7"}}},"metadata":{"id":"uc-demo","last_modified":"2024-07-19T13:20:03.410779+00:00","created":"2024-07-19T13:20:03.410779+00:00"}}
